+++ src/GPW/include/gpw_kernel.h
@@ -9,15 +9,15 @@ double GPWKernel1(const int *configA, const int sizeA, const int *configB,
 
 /* computes the kernel matrix (k(1) kernel) for two lists of configurations
 (in bitstring representation) */
-void GPWKernel1Mat(const int *configsAUp, const int *configsADown,
-                   const int sizeA, const int numA, const int *configsBUp,
-                   const int *configsBDown, const int sizeB, const int numB,
+void GPWKernel1Mat(const unsigned long *configsAUp, const unsigned long *configsADown,
+                   const int sizeA, const int numA, const unsigned long *configsBUp,
+                   const unsigned long *configsBDown, const int sizeB, const int numB,
                    const int tRSym, const int shift, const int symmetric,
                    double *kernelMatr);
 
 /* computes the kernel vector (k(1) kernel) for one reference configuration
 and a list of confgurations in bitstring representation) */
-void GPWKernel1Vec(const int *configsAUp, const int *configsADown,
+void GPWKernel1Vec(const unsigned long *configsAUp, const unsigned long *configsADown,
                    const int sizeA, const int numA, const int *configRef,
                    const int sizeRef, const int tRSym, const int shift,
                    double *kernelVec);
@@ -34,16 +34,16 @@ double GPWKernelN(const int *configA, const int *neighboursA, const int sizeA,
 
 /* computes the kernel matrix (k(n) kernel) for two lists of configurations
 (in bitstring representation) */
-void GPWKernelNMat(const int *configsAUp, const int *configsADown,
+void GPWKernelNMat(const unsigned long *configsAUp, const unsigned long *configsADown,
                    const int *neighboursA, const int sizeA, const int numA,
-                   const int *configsBUp, const int *configsBDown,
+                   const unsigned long *configsBUp, const unsigned long *configsBDown,
                    const int *neighboursB, const int sizeB, const int numB,
                    const int dim, const int n, const int tRSym, const int shift,
                    const int symmetric, double *kernelMatr);
 
 /* computes the kernel vector (k(n) kernel) for one reference configuration
 and a list of confgurations in bitstring representation) */
-void GPWKernelNVec(const int *configsAUp, const int *configsADown,
+void GPWKernelNVec(const unsigned long *configsAUp, const unsigned long *configsADown,
                    const int *neighboursA, const int sizeA, const int numA,
                    const int *configRef, const int *neighboursRef,
                    const int sizeRef, const int dim, const int n,
@@ -67,9 +67,9 @@ double GPWKernel(const int *sysCfg, const int *sysNeighbours, const int sysSize,
 
 /* computes the kernel matrix (complete kernel) for two lists of configurations
 (in bitstring representation) */
-void GPWKernelMat(const int *configsAUp, const int *configsADown,
+void GPWKernelMat(const unsigned long *configsAUp, const unsigned long *configsADown,
                   const int *neighboursA, const int sizeA, const int numA,
-                  const int *configsBUp, const int *configsBDown,
+                  const unsigned long *configsBUp, const unsigned long *configsBDown,
                   const int *neighboursB, const int sizeB, const int numB,
                   const int dim, const int power, const int rC,
                   const double theta0, const double thetaC,
@@ -78,7 +78,7 @@ void GPWKernelMat(const int *configsAUp, const int *configsADown,
 
 /* computes the kernel vector (complete kernel) for one reference configuration
 and a list of confgurations in bitstring representation) */
-void GPWKernelVec(const int *configsAUp, const int *configsADown,
+void GPWKernelVec(const unsigned long *configsAUp, const unsigned long *configsADown,
                   const int *neighboursA, const int sizeA, const int numA,
                   const int *configRef, const int *neighboursRef,
                   const int sizeRef, const int dim, const int power,
